engine/pyproject.toml
[project]
name = "amethyst-engine"
version = "0.1.0"
description = "Amethyst execution engine"
authors = [
    {name = "Arko Bhattacharya",email = "arko@fask.ai"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "a2a-sdk (>=0.2.14,<0.3.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "fastapi (>=0.115.0,<0.116.0)",
    "grpcio (>=1.73.1,<2.0.0)",
    "protobuf (>=6.31.1,<7.0.0)",
    "googleapis-common-protos (>=1.70.0,<2.0.0)",
    "mcp (>=1.1.0,<2.0.0)",
    "openai (>=1.0.0,<2.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
